backend/requirements.txt
fastapi>=0.128.0
uvicorn[standard]>=0.27.1
sqlmodel>=0.0.16
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.9
httpx>=0.28.1
google-genai
openai==1.14.0
pydantic-settings==2.2.1
email-validator>=2.0.0
ruff==0.4.2
black==24.2.0
pre-commit==3.7.0

# Testing & Typing
pytest>=8.1.1
pytest-asyncio>=0.23.6
mypy
types-python-jose
types-passlib
types-ujson
sqlalchemy-stubs
